# center/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.ChangePassword:
    properties:
      id:
        type: integer
      password:
        type: string
    required:
    - password
    type: object
...
  model.CreateUser:
    properties:
      group_id:
        type: integer
      name:
        type: string
      password:
        type: string
    required:
    - group_id
    - name
    - password
    type: object
...
  model.GroupInfo:
    properties:
      created_at:
        type: string
      group_name:
        type: string
      id:
        type: integer
    type: object
...
  model.PageResult:
    properties:
      items: {}
      total:
        type: integer
    type: object
...
  model.RoleInfo:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_name:
        type: string
    type: object
...
  model.UpdateUser:
    properties:
      group_id:
        type: integer
      id:
        type: integer
      name:
        type: string
      valid:
        type: integer
    required:
    - group_id
    - name
    - valid
    type: object
  model.UserInfo:
    properties:
      created_at:
        type: string
      group:
        $ref: '#/definitions/model.GroupInfo'
      id:
        type: integer
      name:
        type: string
      role:
        $ref: '#/definitions/model.RoleInfo'
      valid:
        type: integer
    type: object
  model.ValidUser:
    properties:
      id:
        type: integer
      valid:
        type: integer
    type: object
...
host: http://39.99.155.139:9918
info:
  contact:
    email: support@sensdata.com
    name: API Support
    url: https://static.sensdata.com/support
  description: This is the API documentation for idb application.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://static.sensdata.com/terms/
  title: IDB API Documentation
  version: "1.0"
paths:
...
  /users:
    delete:
      consumes:
      - application/json
      description: 删除用户
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: delete user
      tags:
      - User
    get:
      consumes:
      - application/json
      description: 获取用户列表
      parameters:
      - description: Page number
        in: query
        name: page
        required: true
        type: integer
      - description: Page size
        in: query
        name: page_size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PageResult'
      summary: get user list
      tags:
      - User
    post:
      consumes:
      - application/json
      description: 新增用户
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.UserInfo'
      summary: create user
      tags:
      - User
    put:
      consumes:
      - application/json
      description: 更新用户
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: update user
      tags:
      - User
  /users/password:
    put:
      consumes:
      - application/json
      description: 更新密码
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: update password
      tags:
      - User
...
  /users/valid:
    put:
      consumes:
      - application/json
      description: 禁用/启用用户
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ValidUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: valid user
      tags:
      - User
schemes:
- http
swagger: "2.0"